requirements.txt
streamlit
torch
faiss-cpu
transformers
sentence-transformers
numpy
pypdf
pandas
scikit-learn
gensims